+++ CHANGELOG.md
@@ -9,6 +9,7 @@ BUG FIXES:
 
 FEATURES:
 * clickhouse: add new parameters for user settings, server config.
+* clickhouse: add feature to configure resources for shards.
 * serverless: add `container` support in `yandex_function_trigger` resource and data source
 * vpc: add `deletion_protection` attribute in `yandex_vpc_address` resource and data source
 * vpc: add `deletion_protection` attribute in `yandex_lb_network_load_balancer` resource and data source